+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+## 1.45.0
+
 * fix(addons): parse `maintenance-window-hour` as an int
 * feat(session): remove region cache on logout
 * feat: add JSON output for the `apps` command